Fix clippy lints introduced by newer stable toolchain
manual checked division in queue2/tracer, unnecessary cast in
audio_decoder_element, collapsible_match allow in example 47 (the
suggested match guard cannot take the buffer by value).

+++ src/elements/codec/audio_decoder_element.rs
@@ -140,7 +140,7 @@ impl<D: AudioDecoder + 'static> Transform for AudioDecoderElement<D> {
         // Preserve input metadata and update PTS/duration
         let mut metadata = buffer.metadata().clone();
         metadata.pts = crate::clock::ClockTime::from_nanos(samples.pts as u64);
-        metadata.duration = crate::clock::ClockTime::from_nanos(samples.duration_nanos() as u64);
+        metadata.duration = crate::clock::ClockTime::from_nanos(samples.duration_nanos());
 
         Ok(Output::single(Buffer::new(
             MemoryHandle::with_len(slot, samples.data.len()),
